Refactor: extract BlueprintSchemaPropertyRow from BlueprintSchemaEditor
Per docs/follow-ups-round-2.md section 2d. Structural extraction only —
the single-row visual-mode renderer moves to a sibling file, values and
callbacks pass in as props instead of closures. No behavior change; enum
raw-text and uiMap state stay in the parent and flow down via props.
